Unify the sercom pad macros across samd21 and samd51
